Create Minimise_pso.py
Particle swarm is a population-based algorithm. The python code reads the data from csv file. Initial conditions and constraints are specified in the code. It uses minimise() function from psopy package and uses the constraints to finds the optimum value of every sample present in the csv dataset. The output is generated at the command line as well as pushed into another csv file.